Keep broker native code out of web bundle
